dx1238: rename caseInternal
