+++ shared/src/business/entities/cases/IncompletePaperCase.test.ts
@@ -6,7 +6,7 @@ import {
   PAYMENT_STATUS,
   ROLES,
 } from '../EntityConstants';
-import { CaseInternal } from './CaseInternal';
+import { CaseInternal } from './IncompletePaperCase';
 import { Correspondence } from '../Correspondence';
 import { applicationContext } from '../../test/createTestApplicationContext';
