+++ shared/src/business/entities/cases/IncompletePaperCase.ts
